requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.2
pandas==2.2.1
SQLAlchemy==2.0.27
beautifulsoup4==4.12.3
google-genai